#88 - refactor file names
