+++ src/Http/Requests/Account/UpdateUserDataRequest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Blumilk\Meetup\Core\Http\Requests;
+namespace Blumilk\Meetup\Core\Http\Requests\Account;
 
 use Blumilk\Meetup\Core\Http\Requests\Speaker\Rules\AvatarFileRules;
 use Illuminate\Foundation\Http\FormRequest;
