+++ src/Http/Routing/WebRouting.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Blumilk\Meetup\Core\Http\Routing;
 
+use Blumilk\Meetup\Core\Http\Controllers\AccountController;
 use Blumilk\Meetup\Core\Http\Controllers\Auth\EmailVerificationController;
 use Blumilk\Meetup\Core\Http\Controllers\Auth\LoginController;
 use Blumilk\Meetup\Core\Http\Controllers\Auth\PasswordResetController;
@@ -16,7 +17,6 @@ use Blumilk\Meetup\Core\Http\Controllers\OrganizationController;
 use Blumilk\Meetup\Core\Http\Controllers\OrganizationProfileController;
 use Blumilk\Meetup\Core\Http\Controllers\SpeakersController;
 use Blumilk\Meetup\Core\Http\Controllers\StaticController;
-use Blumilk\Meetup\Core\Http\Controllers\UserController;
 
 class WebRouting extends Routing
 {
@@ -55,7 +55,7 @@ class WebRouting extends Routing
             $this->router->get("/auth/facebook/callback", "handleFacebookCallback");
         });
 
-        $this->router->controller(UserController::class)->middleware("auth")->group(function (): void {
+        $this->router->controller(AccountController::class)->middleware("auth")->group(function (): void {
             $this->router->get("/auth/profile", "index")->name("user.profile");
             $this->router->get("/auth/profile/password", "editPassword")->name("user.profile.password");
             $this->router->get("/auth/profile/edit", "editData")->name("user.profile.edit");
